Move filesystem access check functions outside of the filesystem plugin
